Исправить "operator is not unique" при создании визита
INSERT в finance_entries умножал/вычитал два параметра ($4 * $5) без
приведения типов — Postgres не мог выбрать оператор между двумя
"unknown"-параметрами и падал с operator is not unique: unknown * unknown,
визит не сохранялся. Явные ::numeric убирают неоднозначность.

+++ backend/src/modules/visits/visits.routes.js
@@ -267,7 +267,7 @@ router.post(
       // finance_entries (источник выручки), а не расчёт налету из visits.
       await dbClient.query(
         `INSERT INTO finance_entries (company_id, source, visit_id, membership_id, amount, occurred_at, created_by_user_id)
-         VALUES ($1, 'auto_from_visit', $2, $3, ROUND($4 - ($4 * $5 / 100), 2), $6::timestamptz::date, $7)`,
+         VALUES ($1, 'auto_from_visit', $2, $3, ROUND($4::numeric - ($4::numeric * $5::numeric / 100), 2), $6::timestamptz::date, $7)`,
         [req.tenant.companyId, visitId, resolvedMasterId, amount, discountPercent || 0, insert.rows[0].visit_at, req.user.id]
       );
